New Branch
I have added a new branch named Debugging. The program outputs both time formats, displays the menu options, and loops with user input. The main problem is that the seconds are getting printed twice. Also that the user input is recognized when 1-3 is entered, but it does not display the correct output. It only displays the current time. However, the program does recognize when the user enters 4 or a digit not 1-3 and quits; which is good.