Debugging-File
I added a separate file for bugging. The first thing I notice is that the seconds are repeated, even with the simplified code. I think there might be an error with the .tm_sec. I will do some research to understand what is going on. Also, I had to clear the mainclockcode file in order to have only one file debug.