Part 2 of 1st Bug
The second half of the first bug is fixed. The first half focused on fixing the first print out statement of the clock. The second half fixed the print statement that comes after the user makes a choice. The time isn't being added, but it is at least appearing in the correct format.