All bugs fixed
All bugs have been fixed. The time prints properly and correctly recognizes user input. To fix the bug, I put the time struct in the main instead of in a separate function! I also made some minor changes so that the 12 hour clock only prints 12 hours and the same for the 24 hour clock. My final commit will include comments in the code.